chore: move share-able mocks to utils
