+++ geoprocessing/src/utils/__mocks__/connection.ts
